pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "cryptography>=44.0.2",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "google-generativeai>=0.8.4",
    "oauthlib>=3.2.2",
    "openai>=1.65.4",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "requests>=2.32.3",
    "sendgrid>=6.11.0",
    "slack-sdk>=3.34.0",
    "telethon>=1.39.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.6",
]
